@@ -212,7 +212,7 @@
       },
 
       setPosition(newPosition) {
-        if (newPosition === null) return;
+        if (newPosition === null || isNaN(newPosition)) return;
         if (newPosition < 0) {
           newPosition = 0;
         } else if (newPosition > 100) {
