@@ -124,7 +124,7 @@ export default {
                             : ''
                         }
                         {
-                          column.filterable
+                          (column.filters && column.filters.length) || column.filterMethod
                             ? <span class="el-table__column-filter-trigger" on-click={ ($event) => this.handleFilterClick($event, column) }><i class={ ['el-icon-arrow-down', column.filterOpened ? 'el-icon-arrow-up' : ''] }></i></span>
                             : ''
                         }
